♻️ Refactor: 인풋 컴포넌트 재적용
